add different profile fragment classes
